refactor: 페이지별 route 정보와 라우팅 로직 분리
